Mejora en la configuración de la BD
